Removed code taken from core, using those features instead
