+++ src/de/mod.rs
@@ -1,7 +1,6 @@
 //! Decoder-based structs and traits.
 
 mod decoder;
-mod impl_core;
 mod impl_tuples;
 mod impls;
